Cleanup, proper example
This sets up the control plane with 2 Dockerized WireGuard servers.

On startup, the WireGuard servers POST to the control plane on port
8081 their configuration, which is stored in Redis and passed on to
WireGuards clients as they send a request with their public key to
the control plane. After sanity checks, public key and IP of clients
are then published on a specific Redis channel (currently the
interface name).

The api then uses mullvad/message-queue to subscribe to those channels
and write events to connected WireGuard servers over WebSocket on port
:8080/channel/<INTERFACE_NAME>.

The wss WireGuard servers finally update their interfaces when they
receive ADD or DEL messages over WebSocket from the control plane.

This allows for distributed WireGuard servers, with centralised auth,
and separation of concerns when it comes to auth vs. networking - the
only shared configuration between control plane and WireGuard servers
needed is the interface name (which could be any ID).